+++ source/extensions/omni.isaac.core/omni/isaac/core/tests/simulation_context/test_simulation_context.py
@@ -15,6 +15,7 @@ from omni.isaac.core import World, SimulationContext
 class TestSimulationContext(omni.kit.test.AsyncTestCaseFailOnLogError):
     # Before running each test
     async def setUp(self):
+        World.clear_instance()
         pass
 
     # After running each test