+++ source/extensions/omni.isaac.dynamic_control/premake5.lua
@@ -15,7 +15,12 @@ project_ext_plugin(ext, "omni.isaac.dynamic_control.plugin")
         "%{root}/_build/target-deps/pxshared/include",
         "%{root}/_build/target-deps/nv_usd/%{cfg.buildcfg}/include",
         "%{root}/_build/target-deps/usd_ext_physics/%{cfg.buildcfg}/include",
-        "%{root}/_build/target-deps/omni_physics/include"
+        "%{root}/_build/target-deps/omni_physics/include",
+        "%{root}/_build/target-deps/client_library/include",
+        "%{root}/_build/target-deps/carbonite/include",
+        "%{root}/_build/target-deps/rtx_plugins/include",
+
+
 
      }
      libdirs {
@@ -23,7 +28,7 @@ project_ext_plugin(ext, "omni.isaac.dynamic_control.plugin")
         "%{root}/_build/target-deps/usd_ext_physics/%{cfg.buildcfg}/lib"
     }
 
-    links {"gf", "sdf", "usd", "usdGeom","usdUtils"}
+    links {"gf", "sdf", "usd", "usdGeom","usdUtils", "tf", "arch"}
 
     filter { "system:linux" }
         includedirs {