+++ source/extensions/omni.isaac.dynamic_control/python/tests/test_core.py
@@ -10,6 +10,7 @@
 import omni.kit.test
 import omni.usd
 from omni.isaac.dynamic_control import _dynamic_control
+from pxr import Sdf
 
 # Having a test class dervived from omni.kit.test.AsyncTestCase declared on the root of module will make it auto-discoverable by omni.kit.test
 class TestCore(omni.kit.test.AsyncTestCaseFailOnLogError):
@@ -47,3 +48,23 @@ class TestCore(omni.kit.test.AsyncTestCaseFailOnLogError):
         self.assertEqual("(1, 2, 3), (4, 5, 6)", str(v))
         self.assertEqual("(1, 2, 3), (1, 2, 3, 4), (1, 2, 3), (4, 5, 6)", str(_dynamic_control.RigidBodyState(t, v)))
         self.assertEqual("(1, 2, 3)", str(_dynamic_control.DofState(1, 2, 3)))
+
+    async def test_delete(self):
+        await omni.kit.app.get_app().next_update_async()
+        self._stage = omni.usd.get_context().get_stage()
+        ext_manager = omni.kit.app.get_app().get_extension_manager()
+        ext_id = ext_manager.get_enabled_extension_id("omni.isaac.dynamic_control")
+        self._extension_path = ext_manager.get_extension_path(ext_id)
+        await omni.kit.app.get_app().next_update_async()
+        prim_a = self._stage.DefinePrim("/World/Franka_1", "Xform")
+        prim_a.GetReferences().AddReference(self._extension_path + "/data/usd/robots/franka/franka.usd")
+        prim_b = self._stage.DefinePrim("/World/Franka_2", "Xform")
+        prim_b.GetReferences().AddReference(self._extension_path + "/data/usd/robots/franka/franka.usd")
+        self._timeline.play()
+        await omni.kit.app.get_app().next_update_async()
+        self._handle = self._dc.get_articulation("/World/Franka_1")
+        await omni.kit.app.get_app().next_update_async()
+        with Sdf.ChangeBlock():
+            omni.usd.commands.DeletePrimsCommand(["/World/Franka_1"]).do()
+            omni.usd.commands.DeletePrimsCommand(["/World/Franka_2"]).do()
+        await omni.kit.app.get_app().next_update_async()